Apply layer normalization to the output of each gate in GRU.
